[M453] Fix warnings generated by armcc

+++ targets/cmsis/TARGET_NUVOTON/TARGET_M451/cmsis_nvic.h
@@ -24,12 +24,12 @@
 #define NVIC_NUM_VECTORS     (NVIC_USER_IRQ_OFFSET + NVIC_USER_IRQ_NUMBER)
 
 #if defined(__CC_ARM)
-#   define NVIC_RAM_VECTOR_ADDRESS  &Image$$ER_IRAMVEC$$ZI$$Base
+#   define NVIC_RAM_VECTOR_ADDRESS  ((uint32_t) &Image$$ER_IRAMVEC$$ZI$$Base)
 #elif defined(__ICCARM__)
 #   pragma section = "IRAMVEC"
 #   define NVIC_RAM_VECTOR_ADDRESS  ((uint32_t) __section_begin("IRAMVEC"))
 #elif defined(__GNUC__)
-#   define NVIC_RAM_VECTOR_ADDRESS  &__start_vector_table__
+#   define NVIC_RAM_VECTOR_ADDRESS  ((uint32_t) &__start_vector_table__)
 #endif
